Move StateStoreWaitForFiles to compaction-core
