+++ java/compaction/compaction-core/src/main/java/sleeper/compaction/task/TimedOutWaitingForFileAssignmentsException.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.compaction.job.execution;
+package sleeper.compaction.task;
 
 public class TimedOutWaitingForFileAssignmentsException extends RuntimeException {
