+++ java/compaction/compaction-job-execution/src/test/java/sleeper/compaction/job/execution/StateStoreWaitForFilesTestHelper.java
@@ -15,6 +15,7 @@
  */
 package sleeper.compaction.job.execution;
 
+import sleeper.compaction.task.StateStoreWaitForFiles;
 import sleeper.configuration.properties.table.TablePropertiesProvider;
 import sleeper.core.util.ExponentialBackoffWithJitter;
 import sleeper.statestore.StateStoreProvider;
@@ -27,7 +28,7 @@ public class StateStoreWaitForFilesTestHelper {
     }
 
     public static StateStoreWaitForFiles waitWithRetries(int retries, StateStoreProvider stateStoreProvider, TablePropertiesProvider tablePropertiesProvider) {
-        return new StateStoreWaitForFiles(retries, backoffNoJitter(), stateStoreProvider, tablePropertiesProvider);
+        return new StateStoreWaitForFiles(retries, backoffNoJitter(), stateStoreProvider.byTableId(tablePropertiesProvider));
     }
 
     private static ExponentialBackoffWithJitter backoffNoJitter() {
